fixed cast error in gradeReceivedEmail

+++ backend/email-microservice/src/main/java/edu/oswego/cs/services/EmailService.java
@@ -221,7 +221,7 @@ public class EmailService {
         Document assignment = new AssignmentInterface().getSpecifiedAssignment(courseID, assignmentID);
         Document submission = new AssignmentInterface().getSubmission(assignmentID, teamID);
 
-        if(submission.getInteger("grade") != -1){ //may need to be changed as the grade system is updated.
+        if((int)(submission.getDouble("grade").doubleValue()) != -1){ //may need to be changed as the grade system is updated.
             return;
         }
